Organized the engine to decouple functionality
