+++ src/state/Constants.h
@@ -1,0 +1,8 @@
+#ifndef DB9C663F_7757_4E5F_9CDD_23E9A09D9515
+#define DB9C663F_7757_4E5F_9CDD_23E9A09D9515
+
+#define BOARD_SIZE 64
+#define BOARD_LENGTH 8
+#define MAX_LEGAL_MOVES 218
+
+#endif /* DB9C663F_7757_4E5F_9CDD_23E9A09D9515 */
